requirements.txt
openai==0.27.0
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.1
python-telegram-bot==20.1
